abis jumatan

+++ database/migrations/2022_01_20_094150_create_obat_gudang_table.php
@@ -16,8 +16,8 @@ class CreateObatGudangTable extends Migration
         Schema::create('obat_gudang', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('obat_id');
-            $table->decimal('harga_jual')->default(0);
-            $table->decimal('harga_beli')->default(0);
+            $table->bigInteger('harga_jual')->default(0);
+            $table->bigInteger('harga_beli')->default(0);
             $table->bigInteger('stok')->default(0);
             $table->bigInteger('minimal_stok')->default(0);
             $table->bigInteger('maksimal_stok')->default(0);
